requirements.txt
onnxruntime-gpu>=1.8.1
opencv-python>=4.1.2
tesserocr==2.5.2
numpy>=1.18.5
PyYAML>=5.3.1
tqdm>=4.41.0
Pillow>=7.1.2
torch>=1.7.0
torchvision>=0.8.1
easydict>=1.8
